checkbox, mutator

+++ app/Http/Controllers/Admin/ProductController.php
@@ -44,9 +44,11 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')){
+            $params['image'] = $request->file('image')->store('products');
+        }
         Product::create($params);
         return redirect()->route('products.index');
     }
@@ -83,10 +85,19 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')){
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
+
+        foreach (['hit', 'new', 'recommend'] as $fieldName){
+            if (!isset($params[$fieldName])){
+                $params[$fieldName] = 0;
+            }
+        }
+
         $product->update($params);
         return redirect()->route('products.index');
     }
